requirements.txt
# Data Streaming
kafka-python==2.0.2
confluent-kafka==2.3.0

# Database
clickhouse-driver==0.2.6
clickhouse-connect==0.6.23

# Data Processing
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Online Machine Learning
river==0.21.0

# Dashboard & Visualization
streamlit==1.29.0
plotly==5.18.0
altair==5.2.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Code Quality
black==23.12.1
pylint==3.0.3
mypy==1.7.1

# Logging
loguru==0.7.2
